// Capstone Project/src/app/pages/landlord/house-create-page/screen/types.ts
export enum hostHouseCreateEnum {
  GET_ALL_CITY_REQUEST = 'APP/HOST_HOUSE_CREATE/GET_ALL_CITY_REQUEST',
  GET_ALL_CITY_SUCCESS = 'APP/HOST_HOUSE_CREATE/GET_ALL_CITY_SUCCESS',
  GET_ALL_CITY_FAILURE = 'APP/HOST_HOUSE_CREATE/GET_ALL_CITY_FAILURE',

  GET_DISTRICT_REQUEST = 'APP/HOST_HOUSE_CREATE/GET_DISTRICT_REQUEST',
  GET_DISTRICT_SUCCESS = 'APP/HOST_HOUSE_CREATE/GET_DISTRICT_SUCCESS',
  GET_DISTRICT_FAILURE = 'APP/HOST_HOUSE_CREATE/GET_DISTRICT_FAILURE',

  GET_VILLAGE_REQUEST = 'APP/HOST_HOUSE_CREATE/GET_VILLAGE_REQUEST',
  GET_VILLAGE_SUCCESS = 'APP/HOST_HOUSE_CREATE/GET_VILLAGE_SUCCESS',
  GET_VILLAGE_FAILURE = 'APP/HOST_HOUSE_CREATE/GET_VILLAGE_FAILURE',

  CREATE_HOUSE_REQUEST = 'APP/HOST_HOUSE_CREATE/CREATE_HOUSE_REQUEST',
  CREATE_HOUSE_SUCCESS = 'APP/HOST_HOUSE_CREATE/CREATE_HOUSE_SUCCESS',
  CREATE_HOUSE_FAILURE = 'APP/HOST_HOUSE_CREATE/CREATE_HOUSE_FAILURE',

  BTN_LOADING_STATE = 'APP/HOST/BTN_LOADING_STATE',
  HOST_HOUSE_CREATE_STATE = 'APP/HOST/HOST_HOUSE_CREATE_STATE',
}